+++ packages/sharedServer/src/services/feedItems.server.ts
@@ -6,8 +6,10 @@ import {assertNever, omitUndefined} from '@shared/lib/utils.shared';
 
 import type {AccountId} from '@shared/types/accounts.types';
 import {FeedItemType} from '@shared/types/feedItems.types';
-import type {FeedItem, FeedItemFromStorage, FeedItemId} from '@shared/types/feedItems.types';
+import type {FeedItem, FeedItemId} from '@shared/types/feedItems.types';
 import type {AsyncResult, Result} from '@shared/types/results.types';
+
+import type {FeedItemFromStorage} from '@shared/schemas/feedItem.schema';
 
 import {eventLogService} from '@sharedServer/services/eventLog.server';
 import {storage} from '@sharedServer/services/firebase.server';
